# docs/swagger.yml
openapi: 3.0.0
info:
  title: Wallet Core API
  description: Trust Wallet Core API wrapper and more.
  version: 0.0.0
servers:
  - url: http://127.0.0.1:3000
paths:
  /health:
    get:
      tags:
        - Health
      summary: Health
      description: This endpoint will check the readiness and liveness of the service.
      operationId: Health
      responses:
        204:
          description: No Content response.

  /v1/hdwallet:
    post:
      tags:
        - HDWallet
      summary: Create an HDWallet
      description: By using this endpoint, you can create an HDWallet using request parameters.
      operationId: CreateHDWallet
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateHDWalletRequest'
      responses:
        200:
          description: Response to creating an HDWallet.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateHDWalletResponse'
        400:
          description: Error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /v1/{symbol}/transaction/{transactionId}:
    get:
      tags:
        - Transaction
      summary: Get Transaction Info
      description: By using this endpoint, you can retrieve the transaction info.
      operationId: GetTransactionInfo
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
            example: BTC
        - name: network
          in: query
          required: false
          schema:
            type: string
            example: Bitcoin
        - name: transactionId
          in: path
          required: true
          schema:
            type: string
            example: 'bbd4a06ba81cd97a8f1e7ab4fd9fea9f324f5d2c83c50de2b19d6f3dc7197c4d'
      responses:
        200:
          description: Response to getting the transaction info.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetTransactionInfoResponse'
        400:
          description: Error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /v1/{symbol}/wallet/{walletId}:
    get:
      tags:
        - Wallet
      summary: Get Wallet Info
      description: You can retrieve the wallet info and its last ten transactions using this endpoint.
      operationId: GetWalletInfo
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
            example: ETH
        - name: network
          in: query
          required: false
          schema:
            type: string
            example: Ethereum
        - name: walletId
          in: path
          required: true
          schema:
            type: string
            example: '0xb3aafc07e4ea2a66866b2d2d1a8e28deee400770'
      responses:
        200:
          description: Response to getting the wallet info.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetWalletInfoResponse'
        400:
          description: Error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /v1/transaction/sign:
    post:
      tags:
        - Transaction
      summary: Sign a Transaction
      description: By using this endpoint, you can sign a transaction.
      operationId: SignTransaction
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignTransactionRequest'
      responses:
        200:
          description: Response to signing a transaction.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignTransactionResponse'
        400:
          description: Error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /v1/transaction/broadcast:
    post:
      tags:
        - Transaction
      summary: Broadcast a Transaction
      description: By using this endpoint, you can broadcast a transaction to the network.
      operationId: BroadcastTransaction
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BroadcastTransactionRequest'
      responses:
        204:
          description: Response to broadcasting a transaction.
        400:
          description: Error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        errors:
          type: array
          items:
            type: string
          example:
            - Something Went Wrong!
    CreateHDWalletRequest:
      type: object
      properties:
        symbol:
          type: string
          example: BTC
        network:
          type: string
          example: Bitcoin
        accountIndex:
          type: integer
          format: int64
          default: 0
          minimum: 0
          example: 0
        externalChain:
          type: integer
          format: int64
          default: 0
          minimum: 0
          example: 0
        addressIndex:
          type: integer
          format: int64
          default: 0
          minimum: 0
          example: 1
      required:
        - symbol
    CreateHDWalletResponse:
      type: object
      properties:
        address:
          type: string
          example: '0xb315B05d06f74b27D3869219f41119585Fcb8656'
        publicKey:
          type: string
          example: '04a6f1596c27229b369e92feab4345cabc6b057baadf53211650b27aa7f1b760f9d3f0cb892067c4b9a0aaa7464a9797aa805c77fdca73da49899c295f47355990'
        privateKey:
          type: string
          example: '3e421d9c54e97085b18dd39a9c9eee29eb3f9a289088a1b19ebbdc81e696a28a'
    GetTransactionInfoResponse:
      type: object
      properties:
        confirmed:
          type: boolean
          example: true
        confirmations:
          type: integer
          format: int64
          example: 26
        requireConfirmations:
          type: integer
          format: int64
          example: 10
        url:
          type: string
          example: 'https://blockchair.com/dogecoin/transaction/bbd4a06ba81cd97a8f1e7ab4fd9fea9f324f5d2c83c50de2b19d6f3dc7197c4d'
    GetWalletInfoResponse:
      type: object
      properties:
        balance:
          type: number
          format: double
          example: -41.5
        transactions:
          type: array
          items:
            type: object
            properties:
              transactionHash:
                type: string
                example: '0xd7ce8996bc3b56d3c7e378acbd4529fdaaba87264efc3f128a5d3baa7533ec4d'
              timestamp:
                type: integer
                format: int64
                example: 1689163103
              balanceChange:
                type: number
                format: double
                example: -41.5
    BroadcastTransactionRequest:
      type: object
      properties:
        symbol:
          type: string
          example: DOGE
        network:
          type: string
          example: Dogecoin
        signedTransactionHex:
          type: string
          example: '01000000024f8459c39f1454a035f5dfb5a3080b961dc07b9ee464d7454251e2efd52dc6d8000000006a47304402207ce248afd39ae4f558a7bb50d2fcb0744b22a71b93fb3476add57231e25449040220028d151e6c5a3d24581570d21917c1eaf82c621fac9bcbb62125e44503d92ed8012103204ac306eda7509174a9d2969d68c76c2cf2af61cd489bbd9b5e15c0bc1f702700000000d3c71144b59780191ca6631f428b9e1d1878c4bfb45eaca18c78751e8252936a010000006a473044022050a659cad3814ce51087ef3b73472b35135fddf7976bdbde427dee36400d08b002203c798194be659535f155a4305bbffc1153cc912b8c523420ff1810d482bf3bb3012103204ac306eda7509174a9d2969d68c76c2cf2af61cd489bbd9b5e15c0bc1f70270000000001045e85cc010000001976a914f223f74dc558724d9ac930dcba9df2ac1266e09588ac00000000'
      required:
        - symbol
        - signedTransactionHex
    SignTransactionRequest:
      type: object
      properties:
        symbol:
          type: string
          example: DOGE
        network:
          type: string
          example: Dogecoin
        data:
          type: object
          description: Depends on the type of your coin.
      required:
        - symbol
        - data
    SignTransactionResponse:
      type: object
      properties:
        transactionId:
          type: string
          example: '8ef2c44c9c67dd57e03b63dfec4d95c53405dcc5d8f0e2da6e45888afddb7205'
        signedTransactionHex:
          type: string
          example: '01000000024f8459c39f1454a035f5dfb5a3080b961dc07b9ee464d7454251e2efd52dc6d8000000006a47304402207ce248afd39ae4f558a7bb50d2fcb0744b22a71b93fb3476add57231e25449040220028d151e6c5a3d24581570d21917c1eaf82c621fac9bcbb62125e44503d92ed8012103204ac306eda7509174a9d2969d68c76c2cf2af61cd489bbd9b5e15c0bc1f702700000000d3c71144b59780191ca6631f428b9e1d1878c4bfb45eaca18c78751e8252936a010000006a473044022050a659cad3814ce51087ef3b73472b35135fddf7976bdbde427dee36400d08b002203c798194be659535f155a4305bbffc1153cc912b8c523420ff1810d482bf3bb3012103204ac306eda7509174a9d2969d68c76c2cf2af61cd489bbd9b5e15c0bc1f70270000000001045e85cc010000001976a914f223f74dc558724d9ac930dcba9df2ac1266e09588ac00000000'